refactor: extract log more
