+++ tools/gfx_tools/compose.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import logging
-from logging.config import dictConfig
 from pathlib import Path
 from textwrap import dedent
 from enum import StrEnum
@@ -10,34 +9,20 @@ import typer
 from typer import Argument as Arg, Option as Opt
 
 from compose.Tileset import Tileset
-from compose.log import (
-    log,
-    FailFastHandler,
-    LevelTrackingFilter,
-    LOGGING_CONFIG,
-)
-
-
-# variable for silent script run (no output to terminal)
-run_silent = True
-
-# variable for progress bar support (tqdm module dependency)
-no_tqdm = False
-
-# File name to ignore containing directory
-ignore_file = ".scratch"
-
-
-class ComposingException(Exception):
-    """
-    Base class for all composing exceptions
-    """
+from compose.log import FailFastHandler, log
 
 
 class LogLevel(StrEnum):
     INFO = "INFO"
     WARNING = "WARNING"
     ERROR = "ERROR"
+
+    @classmethod
+    def as_level(cls, value: str) -> int:
+        return getattr(logging, value)
+
+    def set_level(self):
+        log.setLevel(LogLevel.as_level(self))
 
 
 class FeedBack(StrEnum):
@@ -59,9 +44,9 @@ CONTEXT_SETTINGS = {"help_option_names": ["-h", "--help"]}
 app = typer.Typer(context_settings=CONTEXT_SETTINGS)
 
 
-# fmt: off
 @app.command()
 def main(
+    # fmt: off
     source_dir:       Path     = Arg(...,                         help="Tileset source files directory path"),
     output_dir:       Path     = Arg(None,                        help="Output directory path"),
 
@@ -75,8 +60,8 @@ def main(
 
     loglevel:         LogLevel = Opt(LogLevel.WARNING,            help="set verbosity level"),
     feedback:         FeedBack = Opt(FeedBack.SILENT,             help=FEEDBACK_DESC),
-) -> int | ComposingException:
-# fmt: on
+    # fmt: on
+) -> int:
     """
     Merge all tile entries and PNGs in a compositing tileset directory into
     a tile_config.json and tilesheet .png file(s) ready for use in BN.
@@ -93,51 +78,42 @@ def main(
     if output_dir is None:
         output_dir = source_dir
 
-
-    dictConfig(LOGGING_CONFIG)
-    log.setLevel(getattr(logging, loglevel))
-    log_tracker = LevelTrackingFilter()
-    log.addFilter(log_tracker)
-
     if fail_fast:
         failfast_handler = FailFastHandler()
         failfast_handler.setLevel(logging.ERROR)
         log.addHandler(failfast_handler)
 
-    if feedback == "SILENT":
-        global run_silent
-        run_silent = True
+    loglevel.set_level()
 
-    if feedback == "VERBOSE":
-        run_silent = False
-        feedback = setup_progress_bar()  # may fallback to CONCISE
-
-    if feedback == "CONCISE":
-        run_silent = False
-        global no_tqdm
-        no_tqdm = True  # equal to concise display
+    no_tqdm = False
+    match feedback:
+        case FeedBack.SILENT:
+            run_silent = True
+        case FeedBack.CONCISE:
+            run_silent = False
+            no_tqdm = True
+        case FeedBack.VERBOSE:
+            run_silent = False
+            no_tqdm = False
 
     # compose the tileset
-    try:
-        tileset_worker = Tileset(
-            source_dir,
-            output_dir,
-            use_all,
-            obsolete_fillers,
-            palette_copies,
-            palette,
-            format_json,
-            only_json,
-        )
-        tileset_worker.compose(ignore_file=ignore_file)
+    tileset_worker = Tileset(
+        source_dir,
+        output_dir,
+        use_all,
+        obsolete_fillers,
+        palette_copies,
+        palette,
+        format_json,
+        only_json,
+    )
+    tileset_worker.compose(
+        ignore_file=".scratch",
+        run_silent=run_silent,
+        no_tqdm=no_tqdm,
+    )
 
-    except ComposingException as exception:
-        return exception
-
-    if log_tracker.level >= logging.ERROR:
-        return 1
-
-    return 0
+    return log_tracker.level >= logging.ERROR
 
 
 if __name__ == "__main__":
